@@ -8,11 +8,6 @@ import static eu.europeana.entitymanagement.web.xml.model.XmlConstants.NAMESPACE
 import static eu.europeana.entitymanagement.web.xml.model.XmlConstants.NAMESPACE_RDF;
 import static eu.europeana.entitymanagement.web.xml.model.XmlConstants.NAMESPACE_SKOS;
 import static eu.europeana.entitymanagement.web.xml.model.XmlConstants.PREF_LABEL;
-
-import eu.europeana.entitymanagement.definitions.exceptions.EntityModelCreationException;
-import eu.europeana.entitymanagement.definitions.model.Entity;
-import eu.europeana.entitymanagement.utils.EntityObjectFactory;
-import eu.europeana.entitymanagement.vocabulary.EntityTypes;
 import java.util.ArrayList;
 import java.util.List;
 import javax.xml.bind.annotation.XmlAccessType;
@@ -20,6 +15,10 @@ import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlAttribute;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlTransient;
+import eu.europeana.entitymanagement.definitions.exceptions.EntityModelCreationException;
+import eu.europeana.entitymanagement.definitions.model.Entity;
+import eu.europeana.entitymanagement.utils.EntityObjectFactory;
+import eu.europeana.entitymanagement.vocabulary.EntityTypes;
 
 @XmlAccessorType(XmlAccessType.FIELD)
 @XmlTransient
@@ -46,7 +45,7 @@ public abstract class XmlBaseEntityImpl<T extends Entity> {
   private List<String> hiddenLabel;
 
   @XmlElement(namespace = XmlConstants.NAMESPACE_ORE, name = XmlConstants.IS_AGGREGATED_BY)
-  private XmlAggregationImpl isAggregatedBy;
+  private XmlAggregationWrapperImpl isAggregatedBy;
 
   protected XmlBaseEntityImpl() {
     // default constructor
@@ -65,7 +64,7 @@ public abstract class XmlBaseEntityImpl<T extends Entity> {
       this.hiddenLabel = new ArrayList<String>(entity.getHiddenLabel());
     }
     if (entity.getIsAggregatedBy() != null) {
-      this.isAggregatedBy = new XmlAggregationImpl(entity.getIsAggregatedBy());
+      this.isAggregatedBy = new XmlAggregationWrapperImpl(new XmlAggregationImpl(entity.getIsAggregatedBy()));
     }
 
     this.isShownBy = XmlWebResourceWrapper.fromWebResource(entity.getIsShownBy());
