@@ -1,5 +1,6 @@
 package eu.europeana.entitymanagement;
 
+import java.time.Duration;
 import java.util.Arrays;
 import java.util.Set;
 import org.apache.logging.log4j.LogManager;
@@ -16,6 +17,7 @@ import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.context.ConfigurableApplicationContext;
 import eu.europeana.entitymanagement.batch.model.JobType;
 import eu.europeana.entitymanagement.batch.service.BatchEntityUpdateExecutor;
+import eu.europeana.entitymanagement.web.model.ZohoSyncReport;
 import eu.europeana.entitymanagement.web.service.ZohoSyncService;
 
 /**
@@ -55,7 +57,20 @@ public class EntityManagementApp implements CommandLineRunner {
               .web(WebApplicationType.NONE)
               .run(args);
 
-      LOG.info("Batch update execution complete for {}. Stopping application. ", Arrays.toString(args));
+      LOG.info("Batch scheduling execution complete for {} ", Arrays.toString(args));
+      //3hours
+      int minutes = 180;
+      LOG.info("Wait for completion of asynchonuous processing: {}m", minutes);
+      try {
+        Thread.sleep(Duration.ofMinutes(minutes).toMillis());
+      } catch (InterruptedException e) {
+        LOG.error("Cannot complete execution!", e);
+        SpringApplication.exit(context);
+        System.exit(-2);
+      }
+      
+      //TODO: erorneous execution should be indicated with negative codes
+      LOG.info("Stoping application after scheduling batch processing and waiting {}m for processing schedule updates!", minutes);
       System.exit(SpringApplication.exit(context));
     } else {
       LOG.info("No args provided to application. Starting web server");
@@ -72,19 +87,25 @@ public class EntityManagementApp implements CommandLineRunner {
   public void run(String... args) throws Exception {
     if (hasCmdLineParams(args)) {
       Set<String> tasks = Set.of(args);
-      if(tasks.contains(JobType.SCHEDULE_DELETION.value())) {
-        LOG.info("Executing scheduled deletions");
-        batchUpdateExecutor.runScheduledDeprecationsAndDeletions();
-      }
       
+      //first zoho sync as it runs synchronuous operations
       if(tasks.contains(JobType.ZOHO_SYNC.value())) {
         LOG.info("Executing zoho sync");
-        zohoSyncService.synchronizeModifiedZohoOrganizations();
+        ZohoSyncReport zohoSyncReport = zohoSyncService.synchronizeModifiedZohoOrganizations();
+        LOG.info("Synchronization Report: {}", zohoSyncReport.toString());
+      }
+      
+      if(tasks.contains(JobType.SCHEDULE_DELETION.value())) {
+        //run also the deletions called through the API directly
+        LOG.info("Executing scheduled deletions");
+        batchUpdateExecutor.runScheduledDeprecationsAndDeletions();
+        //TODO: should read the number of scheduled deletions and deprecations from the database and write it to the logs
       }
       
       if(tasks.contains(JobType.SCHEDULE_UPDATE.value())) {
         LOG.info("Executing scheduled updates");
         batchUpdateExecutor.runScheduledUpdate();
+        //TODO: should read the number of scheduled deletions and deprecations from the database and write it to the logs
       }
         
     }
